run: add implicit exit-status test
If `run` is invoked with '=N' or '!' as its
first argument, it will check the exit code
of the invoked command and fail if it does
not match the requirement (exact N for =N,
or any nonzero for '!').

Additionally, `run` now logs each command invoked
as well as its entire output. This will be invisible
when tests pass, but will be shown on test failure
and may offer crucial context for someone analyzing
a failed run.

Signed-off-by: Ed Santiago <santiago@redhat.com>