Improve external function calls tests
The 'report correct line on external function calls' test
does not check which cases include the external function in
the stack trace and which exclude it. Improve the test by
marking each case with the expected stack depth.

Aside from making the test more strict, this commit should
not change behavior.